Feature

+++ models/commentModel.js
@@ -1,23 +1,13 @@
 const connection = require('../db/db');
 
-// NOTE: 댓글 검증
-exports.validateComments = async (user_id, comment_id) => {
-  const query = `SELECT * FROM comment WHERE user_id = ? AND id = ?`;
-  try {
-    const [result] = await connection.query(query, [user_id, comment_id]);
-    return result; 
-  } catch (error) {
-    console.error('Error validating comments:', error.message);
-    throw error;
-  }
-};
 
-
+//TODO: JWT
 // NOTE: 게시글 ID로 댓글 조회
 exports.findCommentsByPostId = async (postId) => {
   const query = `
     SELECT 
         comment.id, 
+        comment.user_id,
         comment.comment, 
         user.nickname AS author, 
         comment.date
@@ -42,7 +32,7 @@ exports.findCommentsByPostId = async (postId) => {
     throw error;
   }
 };
-
+//TODO: JWT
 // NOTE: 댓글 작성
 exports.createComment = async (user_id, post_id, comment, date) => {
   const query = `INSERT INTO comment (user_id, post_id, comment, date) VALUES (?, ?, ?, ?)`;
@@ -56,25 +46,26 @@ exports.createComment = async (user_id, post_id, comment, date) => {
   }
 };
 
+//TODO: JWT
 // NOTE: 댓글 수정
-exports.updateComment = async (comment_id, content) => {
-  const query = `UPDATE comment SET comment = ? WHERE id = ?`;
+exports.updateComment = async (content, date, comment_id, user_id) => {
+  const query = `UPDATE comment SET comment = ? , date = ? WHERE id = ? and user_id = ?`;
 
   try {
-    const [result] = await connection.query(query, [content, comment_id]);
+    const [result] = await connection.query(query, [content, date, comment_id, user_id]);
     return result.affectedRows > 0; 
   } catch (error) {
     console.error('Error updating comment:', error.message);
     throw error;
   }
 };
-
+//TODO: JWT
 // NOTE: 댓글 삭제
-exports.deleteComment = async (comment_id) => {
-  const query = 'DELETE FROM comment WHERE id = ?';
+exports.deleteComment = async (comment_id, user_id) => {
+  const query = 'DELETE FROM comment WHERE id = ? and user_id = ?';
 
   try {
-    const [result] = await connection.query(query, [comment_id]);
+    const [result] = await connection.query(query, [comment_id, user_id]);
     return result.affectedRows > 0; 
   } catch (error) {
     console.error('Error deleting comment:', error.message);
